Move OpenCL related source file into a sub-module.
Some tests may be skipped, some files may need rework but looks
promizing
