+++ pyFAI/opencl/OCLFullSplit.py
