+++ pyFAI/opencl/ocl_azim_csr_dis.py
