+++ pyFAI/opencl/ocl_fullSplit_CSR.pyx
