+++ pyFAI/opencl/ocl_hist_pixelsplit.py
