+++ version.py
@@ -53,7 +53,7 @@ from __future__ import absolute_import, print_function, division
 __authors__ = ["Jérôme Kieffer"]
 __license__ = "MIT"
 __copyright__ = "European Synchrotron Radiation Facility, Grenoble, France"
-__date__ = "27/01/2017"
+__date__ = "02/02/2017"
 __status__ = "production"
 __docformat__ = 'restructuredtext'
 __all__ = ["date", "version_info", "strictversion", "hexversion", "debianversion", "calc_hexversion"]
@@ -69,7 +69,7 @@ MAJOR = 0
 MINOR = 14
 MICRO = 0
 RELEV = "dev"  # <16
-SERIAL = 1  # <16
+SERIAL = 2  # <16
 
 date = __date__
 
@@ -90,13 +90,13 @@ if version_info.releaselevel != "final":
 
 def calc_hexversion(major=0, minor=0, micro=0, releaselevel="dev", serial=0):
     """Calculate the hexadecimal version number from the tuple version_info:
-    
+
     :param major: integer
     :param minor: integer
     :param micro: integer
     :param relev: integer or string
     :param serial: integer
-    :return: integerm always increasing with revision numbers  
+    :return: integerm always increasing with revision numbers
     """
     global RELEASE_LEVEL_VALUE
     try:
